+++ src/index.ts
@@ -1,4 +1,4 @@
 export { rootAtom } from './rootAtom';
 export { BootstrapRoot } from './bootstrapRoot';
-export { bootstrappedAtom } from './bootstrappedAtom';
+export { bootstrappedSelector as bootstrappedAtom } from './bootstrappedAtom';
 export { bootstrappedAtomValueHook } from './bootstrappedAtomValueHook';