+++ src/rootAtom.ts
@@ -1,5 +1,8 @@
 import { atom } from 'recoil';
 
+import type { RootAtom } from './util';
+import { attachedSelectorsSymbol } from './util';
+
 // Root atoms need to be fully under recoil-bootstrap's control, since
 // bootstrapped atoms rely closely on the lifecycle of this atom for all of
 // their functionality. Allowing users more control would likely destabilize
@@ -9,9 +12,11 @@ import { atom } from 'recoil';
  * Creates a root atom.
  *
  * @param key The key to assign to the root atom.
- * @returns The root atom to be passed to a corresponding BoostrapRoot
+ * @returns The root atom to be passed to a corresponding BootstrapRoot
  *  component.
  */
 export function rootAtom<T>(key: string) {
-  return atom<T>({ key });
+  const newRoot = atom<T>({ key }) as RootAtom<T>;
+  newRoot[attachedSelectorsSymbol] = [];
+  return newRoot;
 }