+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed bug where bootstrapped atoms can be accessed outside of their bootstrap roots when still accessed inside of the Recoil root.
 - Added JSDoc comments for everything exported from this module.
+- Added support for resetting state when the bootstrap root is unmounted
 
 ## 1.0.1 (06/23/2023)